Bash around at installing Sphinx
I think I have an sphinx tarball stuck in a cache somewhere and now pip
won't download the tarball again.

The docs say that the --download option will download the tarball
always. Maybe we have an old version of pip because of an old version of
virtualenv and that's why this isn't installing properly?